LPD-78240 portal-search-rest-api: Move TextEmbeddingProvider to shared location
https://liferay.atlassian.net/browse/LPD-78240
