+++ modules/apps/portal-search/portal-search-rest-api/src/main/java/com/liferay/portal/search/rest/text/embeddings/configuration/TextEmbeddingProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal.ml.embedding.text;
+package com.liferay.portal.search.rest.text.embeddings.configuration;
 
 import com.liferay.portal.search.rest.dto.v1_0.EmbeddingProviderConfiguration;
 
@@ -15,5 +15,7 @@ public interface TextEmbeddingProvider {
 	public Double[] getEmbedding(
 		EmbeddingProviderConfiguration embeddingProviderConfiguration,
 		String text);
+
+	public String getProviderName();
 
 }
