+++ modules/apps/portal-search/portal-search-rest-api/src/main/resources/com/liferay/portal/search/rest/text/embeddings/configuration/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
